Update to version 2 of the proposal.
We decided to keep the NS on some subset of "future value types". We
also decided to keep the NS on all the Lock-related API, as we know that
this will be a focus area for Swift 4.

There are a few additional rules to cover these cases, as well as an
update to the table to apply them.