Require operators in classes to be "static" or "final class".
This restricts the scope of the proposal at the request of the core
team, allowing the dynamic dispatch problem to be considered at a later
date.

A list of changes to standard library protocols will be added in a
separate PR, so that if the core team decides they should be considered
in their own proposal, they don't block this change.