Proposal: Rationalizing Sequence end-operation names
This proposal suggests renaming or redesigning many APIs which operate on the beginning or end of a Sequence or Collection, with the goal of producing coherently-named method families. In summary:

* `c.drop…` becomes `c.removing…`
* `c.…First/Last(n)` become `c.…Prefix/Suffix(n)`
* `c.starts(with: prefix…)` becomes `c.hasPrefix(prefix…)`
* `c.index(of: elem/where: isCond)` becomes `c.firstIndex(of: elem/where: isCond)`
* `c.prefix(upTo: i)`, `c.prefix(through: i)`, and `c.suffix(from: i)` become `c[..<i]`, `c[...i]`, and `c[i..<]`

To support that last change, this proposal includes new `IncompleteRange` and `IncompleteClosedRange` types and operators to construct them; it is otherwise not additive.