[status] Rename index.xsl to index.xslt, NFC
As discussed in #517, this changes the MIME type sent by GitHub Pages from application/xml to application/xslt+xml, which is the officially registered MIME type for XSLT 2.0.

As of writing, Chrome quietly complains about the MIME type of the stylesheet, even with this change, but it seems there's not much to be done about it. Chrome also requires type="text/xsl" on the xml-stylesheet declaration.