forInterpolation is now stringInterpolationSegment
This is historically precedented and slightly reduces the impact of the change on the standard library, without being actually incorrect. (I didn’t do this in the first place because I started with the “Minimally changed design” alternative, and it clashed with that one.)

Matching changes to the prototype will land as soon as they’re tested.