Non-exhaustive enums: Contrast a custom case with a custom pattern
It's weird to have a custom case, but it's even weirder to have a
custom pattern that you can only use in one place. As Chris pointed
out on the list, the main reason for the restriction on the pattern's
use is that the proposal makes it a /warning/ to leave known cases
unhandled, not an error. Therefore, leave the proposal describing
`unknown case` (or `unknown default`), but mention that we'd probably
switch to `case #unknown` if we were going to treat this as an error.