Non-exhaustive enums: Cut down SE-0192 per core team feedback
The biggest difference is that it will no longer apply to Swift enums
defined outside the stdlib and overlays; as such, it no longer
provides a spelling for declaring frozen enums. The feature previously
known as `unknown case` is now just spelled `unknown` and can match
any pattern containing a non-frozen enum, or even a pattern containing
an "implicitly frozen enum" from a third-party library (to help with
gradual adoption of changes within a project). See the list of changes
from the first proposal for more information.