[Proposal] Literal initialization via coercion proposal
Currently types conforming to literal protocols are type-checked using regular
initializer rules, which means that for expressions like `UInt32(42)` the
type-checker is going to look up a set of available initializer choices and
attempt them one-by-one trying to deduce the best solution.

This is not always a desired behavior when it comes to numeric and
other literals, because it means that argument is going to be type-checked
separately (most likely to some default literal type like `Int`) and passed
to an initializer call. At the same time coercion behavior would treat
the expression above as `42 as UInt32` where `42` is ascribed to be `UInt32`
and constructed without an intermediate type.