[Proposal] New Syntax for Multiple Trailing Closures
Swift currently supports a special syntax for a single trailing closure
which makes it possible to pass a closure as function's final argument
after parentheses as a block without a label.

This is very useful when the closure expression is long.
We propose to extend this functionality to cover multiple closures instead of just one.