[SE-0286] Revise proposal for improved source compatibility
Taking the suggestions from Pavel Yaskevich and Xiaodi Wu, improve the
source compatibility guarantees by performing both a forward and
backward scan in Swift < 6, resolving ambiguities in favor of the
backward scan. When the backward scan is chosen, provide a warning +
Fix-It that rewrites the code to not use a trailing closure, thereby
eliminating the ambiguity.