Update proposal to reflect latest changes to the Task API - @Sendable, etc.
This commit offers to change some of the code blocks which still had some of the older and deprecated features like `@concurrent`. More generally, though, I changed the function signatures so that they reflect how they are currently in the Task.swift, TaskGroup.swift and TaskCancellation.swift files, and also some of the implementations of the static properties like `currentPriority` which now use the `withUnsafeCurrentTask` function.

For the "free" (excuse my lack of a better term) functions like `detach` and `withTaskGroup` which were previously static but not anymore, I chose to explicitly make them "public" in the document to emphasise that these are to be used "in the wild".

Now, this is truly a nitpick, but I feel like calling static properties like `isCancelled` "functions" is unnatural and uncommon, which is why I changed them to "properties". However, if you have your own reasons for this, I completely understand, and would be happy to revert those changes.