Remove postbuild commands again after discussion, since in the current proposal they do not address all the things that people want from commands that run after both building, testing, etc are complete (i.e. true "postbuild" commands).
Instead add an accessor for the built-products directory and a description of how regular build commands can be configured to act on artifacts after they are produced by adding the path of the artifact as an input to a regular build command.  This is most likely to be useful for custom build tool plugins associated with a root-level package, but can be useful in some cases.