[SE-0313] Post-review revisions (#1363)
* [SE-0313] Remove closure isolation control.

Actor isolation control, specifically `isolated` capture of parameters,
has generated a number of concerns. Subset it out of the proposal; a
separate proposal can bring it back later if it's important.

* [SE-0313] Prohibit multiple isolated parameters.

Based on discussion on the review thread, prohibit multiple
isolated parameters for now. Move that discussion to Future Directions.