Run part of async main synchronously
I'm proposing that we don't place the whole main function on the task
queue, but instead run everything from the start of the program to the
first suspension point synchronously so that programmers may ensure that
initialization code is run before the runloop is running.

Additionally, I'd like to propose that the main function is implicitly
annotated with the @MainActor annotation and that it inherit the
priority of the main thread.