Remove import from compatible declarations
The import declaration is considerably more complicated, and the Core Team has
indicated they would rather such a feature be implemented using novel syntax if
at all.

Actors have been added to the declaration list. As they cannot be
subclassed, they are functionally equivalent to structs for the purpose
of this proposal.