[SE-0341] Restrict opaque type parameters in consuming positions.
Per discussion on the review thread, restrict the use of opaque type
parameters in "consuming" positions. They are not useful here (because
it's hard to create values to pass when you don't have access to the
type), and this follows the reasoning in SE-0328 for the same
restriction. Add a Future Directions section to talk about "flipping"
the generics for opaque types in consuming position.