Proposal: Opaque result types with limited availability
The same-type `return` requirement is unnecessarily strict
when it comes to availability conditions. SE-0244 states
that it should be possible to change the underlying type
in the future version of the library, but that would only
work with pre-existing types. In other words, the same-type
condition does not have to apply across executions of the
same program, the same way that `Hashable` must produce the
same output for the same value during program execution, but
may produce a different value in the next execution.

`#available` is special because it's a checkable form of that:
dynamic availability will not change while the program is
running, but may be different the next time the program runs.