More precise language about future ownership modifiers (#1731)
A callee receives a `shared` parameter by borrow, but whether it's really
passed by borrow in the caller is currently ambiguous, since we will
normally copy arguments in order to minimize access durations on the
original storage. Receiving by borrow is the more interesting interaction
with `move` so focus on that.