Implementing (parts of) the Swift compiler in Swift isn't so ridiculous now.
Remove the item that states that we won't rewrite the Swift compiler in
Swift. While it is still not an explicit goal, we have already started
to reimplement specific parts of the Swift compiler and toolchain in
Swift, and are looking to [incorporate it further]
(e.g.,
https://forums.swift.org/t/implementing-parts-of-the-swift-compiler-in-swift/59524),
so it doesn't make sense for this page to dismiss the idea.