Clarify the scope of the evolution process.
The old wording here is basically the scope of the Language Workgroup,
and we've previously agreed that evolution covers more than that.  This
idea of the core tools necessary to build Swift programs captures the
additional scope as I see it.

I've also tried to clarify the wording around implementation vs. design
without making it too stodgy.  I don't know if I succeeded at that.