Initial proposal for Swift Backtracing API
This SE proposal specifies a straightforward API that allows user code to
obtain and symbolicate backtraces.  It is not intended to be used from
signal handlers or to implement a crash catcher (the runtime already
includes code for that), but may be useful for capturing a backtrace
during normal execution, for instance while debugging a problem with some
code.