Simplify the type inference semantics for Swift 5 compatibility
The way in which Swift 5 compatibility for type inference behavior was
written, based on only having `try` operations, was confusing and
indirect. Instead, use a simpler rule: for the purpose of type
inference in Swift 5, `throw` statements always throw a value of type
`any Error`. This matches the implementation and is easier to
describe.