Process: Suggest refining the proposal in a draft pull request
The current instructions suggest that one should paste a formal proposal
on the forums, repeatedly edit this post, and open a pull request only
once the proposal is sufficiently refined for official review. This is a
common course of action among non-Apple folks, and is not exactly in the
spirit of modern open source development. Since the pull request is part
of the process anyway, suggest starting off with a draft pull request to
enable

* everyone to participate in code review workflows
* changes to be tracked the proper way

throughout the forum discussion.

While we're here, clarify where to place the proposal in the
repository, and that the pull request is to be marked as ready for
review only once the proposal is deemed complete.