Process: Integrate "Preparing an implementation" section into "How to propose a change" list
This small section is perfectly fit to be part of the "Develop the
proposal" step towards officially proposing a change. Besides, we were
already noting the requirement and importance of prototyping an
implementation in this step. Since multiple pull requests are now
mentioned in this step, additionally clarify which one is being referred
to in the subsequent "Request a review" step.