Typed throws proposal (#2184)
* Typed throws proposal

* Update metadata for the typed throws proposal

* Simplify the type inference semantics for Swift 5 compatibility

The way in which Swift 5 compatibility for type inference behavior was
written, based on only having `try` operations, was confusing and
indirect. Instead, use a simpler rule: for the purpose of type
inference in Swift 5, `throw` statements always throw a value of type
`any Error`. This matches the implementation and is easier to
describe.

* Fix a paragraph that trailed off, and describe how the stdlib change works

* Strengthen messaging against the use of typed throws everywhere

* Strengthen introduction's stance on when to use typed throws

* Rewrite introduction and motivation

* Re-incorporate the standard library updates for `rethrows` functions

* Minor cleanups

* Clarify that the requirement on thrown error types is "convertible to any Error"

* Fix grammar

* Revert "Clarify that the requirement on thrown error types is "convertible to any Error""

This reverts commit 47a1ea005a96b4fa7c21f3e2a7c8467c8616d470. We actually
need the stronger requirement.

* Explain why `throws(any Error & Codable)` won't work

* troduce a compatibility feature to rethrows functions to cope with their callees moving to typed throws.

* Tighten wording for the rethrows exception a bit

* Remove "refinement" terminology and discussion, fix examples

The section on "Protocol conformances and refinements" had some
ill-formed code (with a thrown error type that doesn't conform to
Error), and wasn't quite correct about the meaning of protocol
refinements. Simplify it down to just talk about protocol
conformances, which is the important part here.

* Replace the confusing "GenericError" with "SimpleError"

* Note that "When to use typed throws" will go into the API design guidelines

* Update and rename nnnn-typed-throws.md to 0413-typed-throws.md

Updates in advance of review; assigning SE-0413.

---------

Co-authored-by: Stephen Canon <stephentyrone@gmail.com>