Normalize Upcoming Feature Flag for SE-0413
The Upcoming Feature Flag header field should typically only include
the feature flag itself, not additional commentary.

It is usually difficult to explain the exact effect of a flag
in a brief note in this field. The effect of the flag should be
described in the Source Compatibility section, which the
proposal already does.

Also, including additional info in this header field makes it more
difficult for automated tools to parse for the Swift Evolution
Dashboard.