Process: Improvements to "How to propose a change" section (#2201)
* Process: Link "review period" to appropriate section instead of status page

* Process: Reformat "How to propose a change" section

* Make each title in the list stand out by separating it from the
  description with an empty line instead of a colon.
* There is an order to proposing a change. Switch to a numbered list.
* The last point is not itself a step and applies to more than one step.
  Take it out of the list and wrap it in an alert.

* Process: Integrate "Preparing an implementation" section into "How to propose a change" list

This small section is perfectly fit to be part of the "Develop the
proposal" step towards officially proposing a change. Besides, we were
already noting the requirement and importance of prototyping an
implementation in this step. Since multiple pull requests are now
mentioned in this step, additionally clarify which one is being referred
to in the subsequent "Request a review" step.

* Process: Remove a needless repetition

This bit of information does not play a important role here and is
already provided at the start of the next section.