[BitwiseCopyable] Don't infer for exported types.
Types exported from modules built with library evolution may be trivial
during one build and subsequently cease to be trivial.  That means the
compiler must not infer the conformance for types exported from
resilient modules.  To avoid dialecticization based on whether a module
is built in that way, require the annotation for all exported types.