`@isolated(any)` function types (#2357)
* Initial draft of the `@isolated(any)` proposal.

* Guarantee that tasks will start on their formal isolation when the
task function is explicitly isolated to an actor (or is not a closure).

* Add a title, and try to get the multi-paragraph footnote thing to work.

* Link the @isolated(any) proposal to two of the more fundamental
PRs implementing the feature.

* Link to pitch.

* Incorporate Konrad's feedback.

* Add some recommendations about when to use `@isolated(any)`.

* Revise the isolated(any) proposal to soften the dependency on
closure isolation controls, which will not be finished before we
review this proposal.

* Incorporate Doug's feedback.

* Update the isolation rules to be more explicit about the relationship
with the closure isolation pitch.  We're considering this proposal
knowing that some things can't be expressed yet without that pitch.
As long as we add something vaguely like that pitch in the future,
this is how the rules for isolated(any) should apply to it.