SE-0432: Respond to review feedback.
Subset out `borrowing` bindings, and mention them as a future direction in
tandem with `borrowing` bindings appearing in other places in the language.
Make it clear that using a dynamic cast pattern would require the switch
to already be consuming the subject, rather than having the effect of
turning the entire switch into a borrowing switch implicitly.