Rename Array to ContiguousArray.
The examples all refer to a `span()` function. This is unimplementable for
`Array` because it might not have contiguous storage. An `Array` span requires a
`_read` accessor because we need to create a temporary copy of the Array when it
is discontiguous. But we can't talk about `_read` here because it hasn't been
proposed yet.