[SE-0413] Shrink proposal to what was implemented in Swift 6.0
Error type inference for closures did not make it into Swift 6.0,
although the rest of this proposal did (as well as some of the future
directions, e.g., for `AsyncSequence`). Move this inference, and the
other changes behind the upcoming flag `FullTypedThrows`, out to
"Future Directions" and mark this proposal as (otherwise) implemented.