SE-0458: Restructure for improved readability
The Detailed Design was becoming a grab-bag of sections, so give it more
structure. First, we describe all of the sources of memory unsafety in
Swift. Then, note the role of the @safe attribute. Finally, describe
all of the ways in which one can acknowledge unsafety.

One drive-by fix here is to document the unsafe conformances of the
`UnsafeBufferPointer` family of types to the `Collection` protocol
hierarchy.