Replace two outdated links in 'process.md'
This is a small PR to replace two links in `process.md` which still reference the Apple GitHub organization instead of swiftlang. (I noticed these while working on #2729 but they were out of scope for that PR.) The existing links aren't broken, they do redirect successfully, but it's nicer to use the correct URLs.