[SE-0470] Reinstate sendable-metatype model for type checking generics
Based on discussions in the review thread, reinstate the
sendable-metatype model for type checking generic definitions rather
than the potentially-isolated-conformance model. While slightly less
expressive, the sendable-metatype model is easier to reason about and
teach.

It's still possible to generalize the model toward the
potentially-isolated-conformance model, as is already captured in
Future Directions.