[SE-0470] Tighten up requirements within generic functions
Close a potential data race with my attempt to loosen the rules. In
essence, if we're allowed to pass a metatype `T.Type` across isolation
boundaries when there are no constraints of the form `T: P`, then we
will incorrectly allow a dynamic cast `as? any T & P` to bind to an
isolated conformance (because `T` is not `SendableMetatype`).