Update the SE-0443 proposal to reflect the style chosen for diagnostic group names.
The text of SE-0443 uses snake case diagnostic group names, but in practice
we've settled on using Pascal case for diagnostic group names. Additionally,
the name `DeprecatedDeclaration` was chosen for the deprecation diagnostic
group. I think it would be best for the proposal to reflect both of these
stylistic choices so that it can serve as accurate documentation for the
feature in practice.