[SE-0471] Make diagram readable in dark mode
This change replaces the diagram embedded in "SE-0471: Improved Custom SerialExecutor isolation checking for Concurrency Runtime" with a version that is readable in dark mode. I did this by exporting the diagram from OmniGraffle with a solid white background. The previous version had a transparent background, which made the black arrows between the boxes invisible in dark mode.