[SE-0466] Disable `@MainActor` inference when type conforms to a SendableMetatype protocol
Libraries need a way to communicate that certain protocols can't
reasonably be used with global-actor-isolated types. When the primary
definition of a type conforms to such a protocol, it will disable
`@MainActor` inference for that type. This has the effect of keeping
more types `nonisolated` when there's signal that they should not be
on the main actor. There are some protocols like this in the standard
library (e.g., CodingKey, which was hardcoded to be nonisolated in the
synthesis code), where the following code is ill-formed in with
main-actor isolation by default:

    struct S: Codable {
      var a: Int

      // error: CodingKeys inferred to `@MainActor`, which makes the conformance
      // to CodingKey main-actor-isolated and then fails (because the requirements
      // need to be nonisolated).
      enum CodingKeys: CodingKey {
        case a
      }
    }

With this amendment, the conformance to `CodingKey` (which inherits from
`Sendable`), prevents `CodingKeys` from being inferred to be`@MainActor`.