[SE-0470] Infer `nonisolated` on conformances (#2893)
* [SE-0470] Infer `nonisolated` on conformances

Introduce two inference rules that infer `nonisolated` when
`InferIsolatedConformances` is enabled. These inference rules ensure
that most conformances that should be nonisolated remain nonisolated,
particular when we're in the main-actor-by-default mode. The end
result is that much less code needs to change when enabling isolated
conformance inference (including via main-actor-by-default mode). The
two rules are:

* If the protocol inherits from `SendableMetatype` (including
indirectly, e.g., from `Sendable`), then the isolated conformance
could never be used, so it is inferred to be `nonisolated`.
* If all of the declarations used to satisfy protocol requirements are
`nonisolated`, the conformance will be assumed to be `nonisolated`.

* More justification for why this feature is needed, and a possible future removing it

* Fix typo

* Clarify that a nonisolated conformance might evolve to an isolated one

* Update headers and revision history

---------

Co-authored-by: Xiaodi Wu <13952+xwu@users.noreply.github.com>